--- app/src/main/java/com/mtlepberghenov/internship_playground/model/MainModel.java
@@ -1,10 +1,0 @@
-package com.mtlepberghenov.internship_playground.model;
-
-import com.mtlepberghenov.internship_playground.model.entity.Vehicle;
-import io.reactivex.Single;
-import java.util.List;
-
-public interface MainModel {
-
-  Single<List<Vehicle>> getData();
-}
