+++ app/src/main/java/com/mtlepberghenov/internship_playground/screens/vehiclelist/model/VehicleListModel.java
@@ -1,0 +1,10 @@
+package com.mtlepberghenov.internship_playground.screens.vehiclelist.model;
+
+import com.mtlepberghenov.internship_playground.screens.vehiclelist.model.entity.Vehicle;
+import io.reactivex.Single;
+import java.util.List;
+
+public interface VehicleListModel {
+
+  Single<List<Vehicle>> getData();
+}
