+++ app/src/main/java/com/mtlepberghenov/internship_playground/screens/vehiclelist/presenter/VehicleListPresenterImpl.java
@@ -1,11 +1,10 @@
 package com.mtlepberghenov.internship_playground.screens.vehiclelist.presenter;
 
 import com.mtlepberghenov.internship_playground.base.BasePresenter;
-import com.mtlepberghenov.internship_playground.model.MainModel;
-import com.mtlepberghenov.internship_playground.model.entity.Vehicle;
+import com.mtlepberghenov.internship_playground.screens.vehiclelist.model.VehicleListModel;
+import com.mtlepberghenov.internship_playground.screens.vehiclelist.model.entity.Vehicle;
 import com.mtlepberghenov.internship_playground.screens.vehiclelist.view.VehicleListView;
 import com.mtlepberghenov.internship_playground.utils.TextWrapper;
-import io.reactivex.Observer;
 import io.reactivex.SingleObserver;
 import io.reactivex.disposables.Disposable;
 import java.util.List;
@@ -13,12 +12,12 @@ import java.util.List;
 public class VehicleListPresenterImpl extends BasePresenter<VehicleListView>
     implements VehicleListPresenter {
 
-  private MainModel model;
+  private VehicleListModel model;
 
   private static final String ERROR = TextWrapper.newInstance().getErrorMessage();
   public static final String DONE = TextWrapper.newInstance().getDoneMessage();
 
-  public VehicleListPresenterImpl(MainModel model) {
+  public VehicleListPresenterImpl(VehicleListModel model) {
 
     this.model = model;
   }
