+++ app/src/main/java/com/mtlepberghenov/internship_playground/screens/vehiclelist/view/VehicleListView.java
@@ -1,7 +1,7 @@
 package com.mtlepberghenov.internship_playground.screens.vehiclelist.view;
 
 import com.mtlepberghenov.internship_playground.base.BaseView;
-import com.mtlepberghenov.internship_playground.model.entity.Vehicle;
+import com.mtlepberghenov.internship_playground.screens.vehiclelist.model.entity.Vehicle;
 import java.util.List;
 
 public interface VehicleListView extends BaseView {
